+++ modules/splashsync/src/Objects/Order.php
@@ -162,7 +162,7 @@ class Order extends AbstractObject
         $this->spl = Local::getLocalModule();
         //====================================================================//
         // Load Default Currency
-        $this->Currency = new Currency(Configuration::get('PS_CURRENCY_DEFAULT'));
+        $this->Currency = new Currency((int) Configuration::get('PS_CURRENCY_DEFAULT'));
     }
 
     /**
