+++ modules/splashsync/src/Objects/Product.php
@@ -145,6 +145,6 @@ class Product extends AbstractObject
         $this->spl = Local::getLocalModule();
         //====================================================================//
         // Load Default Currency
-        $this->Currency = new Currency(Configuration::get('PS_CURRENCY_DEFAULT'));
+        $this->Currency = new Currency((int) Configuration::get('PS_CURRENCY_DEFAULT'));
     }
 }
